refactor: changes description in line below
- Changing path structure to have lesser folders
- Adding new enviroment variables
