+++ app/commons/enviroment.go
@@ -1,5 +1,10 @@
 package commons
 
+import (
+	"strconv"
+	"time"
+)
+
 func GetMongoHost() string {
 	return GetEnv("MONGO_HOST", "localhost")
 }
@@ -22,4 +27,18 @@ func GetMongoDBName() string {
 
 func GetMongoContainerName() string {
 	return GetEnv("MONGO_CONTAINER_NAME", "mongoservice")
+}
+
+func GetSWAPIURL() string {
+	return GetEnv("SWAPI_URL", "https://swapi.dev/api/")
+}
+
+func GetDefaultTimeout() time.Duration {
+	const defaultTimeout = 100 * time.Second
+	result, err := strconv.Atoi(GetEnv("DEFAULT_TIMEOUT", defaultTimeout.String()))
+	if err != nil {
+		return defaultTimeout
+	}
+
+	return time.Duration(result)
 }
