+++ app/planets/structs/mongo.go
@@ -1,4 +1,4 @@
-package storage
+package structs
 
 import "time"
